refactor(runtimed): split peer runtime sync module
